+++ packages/graph/src/layers/TooltipLayer.tsx
@@ -1,18 +1,39 @@
 import * as React from 'react';
 
-import { Margin, FieldSelector, HoveringState, HoveredPointState } from '../common/types';
+import { Margin, HoveringState, HoveredPointState } from '../common/types';
 import { Tooltip } from '../tooltip/Tooltip';
 import { TooltipItem } from '../tooltip/TooltipItem';
+export interface GroupedY {
+  /** Index of `dataGroups` */
+  groupIdx: number;
+
+  /** Original value on Y */
+  yStrVal: number;
+
+  /** Projected position of Y */
+  yPos: number;
+
+  /** Color string of the point */
+  color: string;
+}
+export interface AxisProjectedValue {
+  /** Projected position of X */
+  xPos: number;
+
+  /** Original value on X */
+  xStrVal: number;
+
+  /** Corresponding data in `dataGroups` */
+  groupedY: GroupedY[];
+}
 
 export interface TooltipLayerProps {
   hovering: HoveringState;
   hoveredPoint: HoveredPointState;
-  data: object[];
+  axisProjectedValues: AxisProjectedValue[];
   graphWidth: number;
   graphHeight: number;
   margin: Margin;
-  xSelector: FieldSelector;
-  ySelector: FieldSelector;
 
   /** X position of the tooltip */
   x?: number;
@@ -20,24 +41,34 @@ export interface TooltipLayerProps {
   /** Y position of the tooltip */
   y?: number;
 
-  getColor: FieldSelector['getScaledVal'];
+  /** X offset for the position of the tooltip */
+  xOffset?: number;
+
+  /** Y offset for the position of the tooltip */
+  yOffset?: number;
 }
 
 /** Generates the tooltip box */
 export const TooltipLayer = ({
   hovering,
   hoveredPoint,
-  data,
+  axisProjectedValues,
   graphWidth,
   graphHeight,
   margin,
-  xSelector,
-  ySelector,
-  x = xSelector.getScaledVal(data[hoveredPoint.index]),
+  x = axisProjectedValues[hoveredPoint.index].xPos,
   y = hoveredPoint.position.y,
-  getColor,
+  xOffset = 0,
+  yOffset = 0,
 }: TooltipLayerProps) => {
-  const { index } = hoveredPoint;
+  const projected = axisProjectedValues[hoveredPoint.index];
+  const tooltipItems = projected.groupedY.map(pointY => (
+    <TooltipItem
+      key={`t-${pointY.yStrVal}`}
+      color={pointY.color}
+      text={`${pointY.yStrVal}`}
+    />
+  ));
 
   return (
     <Tooltip
@@ -45,16 +76,13 @@ export const TooltipLayer = ({
       graphHeight={graphHeight}
       graphMargin={margin}
       position={{
-        x,
-        y,
+        x: x + xOffset,
+        y: y + yOffset,
       }}
       show={hovering}
     >
-      <h3>{xSelector.getFormattedStringVal(data[index])}</h3>
-      <TooltipItem
-        color={getColor(data[index])}
-        text={ySelector.getFormattedStringVal(data[index])}
-      />
+      <h3>{`${projected.xStrVal}`}</h3>
+      {tooltipItems}
     </Tooltip>
   );
 };